+++ src/Schema/Types/LaravelEnumType.php
@@ -108,6 +108,8 @@ class LaravelEnumType extends EnumType
     }
 
     /**
+     * TODO remove check and inline when requiring bensampo/laravel-enum:6
+     *
      * @param  class-string<\BenSampo\Enum\Enum>  $enumClass
      */
     protected function enumDescription(string $enumClass): ?string