+++ tests/Unit/Schema/Types/LaravelEnumTypeTest.php
@@ -38,7 +38,12 @@ final class LaravelEnumTypeTest extends TestCase
     {
         $enumType = new LaravelEnumType(LocalizedUserType::class);
 
-        $this->assertSame('Localized user type', $enumType->config['description']);
+        // TODO remove check when requiring bensampo/laravel-enum:6
+        if (method_exists(LocalizedUserType::class, 'getClassDescription')) {
+            $this->assertSame('Localized user type', $enumType->config['description']);
+        } else {
+            $this->assertSame(null, $enumType->config['description']);
+        }
     }
 
     public function testCustomDescription(): void
@@ -55,7 +60,9 @@ final class LaravelEnumTypeTest extends TestCase
     {
         $enumType = new LaravelEnumType(PartiallyDeprecated::class);
 
-        $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
+        // TODO remove check when requiring bensampo/laravel-enum:6
+        if (method_exists(LocalizedUserType::class, 'getClassDescription')) {
+            $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
 """Partially deprecated"""
 enum PartiallyDeprecated {
   """Not"""
@@ -67,10 +74,24 @@ enum PartiallyDeprecated {
   """Deprecated with reason"""
   DEPRECATED_WITH_REASON @deprecated(reason: "some reason")
 }
-GRAPHQL
-            ,
-            SchemaPrinter::printType($enumType)
-        );
+GRAPHQL,
+                SchemaPrinter::printType($enumType));
+        } else {
+            $this->assertSame(/** @lang GraphQL */ <<<GRAPHQL
+enum PartiallyDeprecated {
+  """Not"""
+  NOT
+
+  """Deprecated"""
+  DEPRECATED @deprecated
+
+  """Deprecated with reason"""
+  DEPRECATED_WITH_REASON @deprecated(reason: "some reason")
+}
+GRAPHQL,
+                SchemaPrinter::printType($enumType));
+        }
+
     }
 
     public function testReceivesEnumInstanceInternally(): void